Added support for priorities and extended area
If tags are found for High, Mediom or Low in the todo, they are removed
and priorities (A), (B) or (C) are added.

If an area isn't found for a todo, then its project is search too for a
todo.

An example of a path for the standard dropbox location was added.